refactor(adr-004): step 6 — split graphql_api/models/ into _base/_interfaces/_infra
The 27 modules under graphql_api/models/ were a flat namespace mixing
storage-aligned base types, cross-cutting interfaces, scalars, and
domain types. Pulling the infrastructure out into private subdirs
makes the domain types visually dominant at the root.

Layout:
  models/
    _base/         # thing, file, table, object_identity  (storage-aligned)
    _interfaces/   # file_interface, inversion_solution_interface,
                   # predecessors_interface, predecessor
    _infra/        # common (scalars), page_info, _dispatch,
                   # inversion_solution_union
    aggregate_inversion_solution.py, automation_task.py,
    general_task.py, inversion_solution.py, inversion_solution_nrml.py,
    openquake_hazard_*.py, relations.py, rupture_set.py,
    scaled_inversion_solution.py, sms_file.py,
    strong_motion_station.py, time_dependent_inversion_solution.py

Mechanics:
  - git mv preserves blame; 12 files moved
  - 75 relative imports (from .X import) converted to absolute
    (from graphql_api.models.<new-path>.X import) via Python script
  - 1 test (test_rupture_set_upload.py) had an inline
    `from graphql_api.models import file_interface as fi` fixed manually
  - CLAUDE.md layered-structure diagram updated

No external impact:
  - GraphQL SDL / wire format unchanged (strawberry types resolve
    identically regardless of Python module location)
  - All clients (weka, runzi, toshi-ui) untouched
  - strawberry.lazy("...") forward-ref strings and importlib
    dispatch-table strings rewritten as part of the import sweep

Verified: 374 passed, 1 skipped (matches the baseline before this
sub-step).

Refs ADR-004 / #319.
