+++ graphql_api/models/_interfaces/file_interface.py
@@ -7,7 +7,7 @@ from strawberry.types import Info
 
 from graphql_api.data.s3 import presigned_download_url
 
-from .common import BigInt, DateTime, JSONString, KeyValuePair
+from graphql_api.models._infra.common import BigInt, DateTime, JSONString, KeyValuePair
 
 
 @strawberry.interface
