+++ graphql_api/models/rupture_set.py
@@ -18,9 +18,9 @@ from graphql_api.data.dynamo import create_file, get_file, get_thing, list_files
 from graphql_api.data.models import RuptureSetData
 from graphql_api.data.s3 import presigned_post_for_file
 
-from .automation_task import RuptureGenerationTask  # noqa: F401 — re-exported for schema.py
-from .common import BigInt, DateTime, KeyValuePair, KeyValuePairInput, client_mutation_id_input_field
-from .file_interface import FileInterface
+from graphql_api.models.automation_task import RuptureGenerationTask  # noqa: F401 — re-exported for schema.py
+from graphql_api.models._infra.common import BigInt, DateTime, KeyValuePair, KeyValuePairInput, client_mutation_id_input_field
+from graphql_api.models._interfaces.file_interface import FileInterface
 
 # ── RuptureSet ────────────────────────────────────────────────────────────────
